Fix license.
I intended to use the MIT license not realizing that I accidentally used the MIT-0 license.
Because the alea and pcg32 algorithms require the licenses to be retained, I don't think I can actually license this under MIT-0.
So, changing to ISC which requires license notice and to match scalene-binary.